requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.2
flask-cors==6.0.1
geographiclib==2.1
geopy==2.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.3.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==22.0.0
